requirements.txt
fastapi
uvicorn
pydantic
requests
google-genai
python-dotenv